requirements.txt
# Streamlit UI
streamlit>=1.32.0

# LangChain + LangGraph + Azure OpenAI
langchain==0.3.27
langchain-openai==0.3.33
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph==0.6.7
langgraph-checkpoint==2.1.1
openai>=1.30.0

# Azure AD authentication (no API key — uses DefaultAzureCredential)
azure-identity==1.25.0

# Tracking and platform — optional at local-dev time (chain.py guards the import).
# Required when deploying the GraphResponsesAgent to Domino/Databricks.
mlflow[databricks]==3.7.0
# aice-mcflow-plugins==0.1.3   # enable if your Domino/AICE environment requires it

# PDF extraction
pdfplumber>=0.11.0
pymupdf>=1.24.0
camelot-py[cv]>=0.11.0

# Word document extraction
python-docx>=1.1.0

# Excel extraction and export
openpyxl>=3.1.0
pandas>=2.2.0
tabulate>=0.9.0

# Vector store (embeddings come from OpenAI API — see doc_qa/retrieval/vectorstore.py)
faiss-cpu>=1.8.0

# Metadata and persistence
python-dotenv>=1.0.0

# Utilities
numpy>=1.26.0
Pillow>=10.0.0
